fiks admin

+++ api/database/migrations/2024_02_13_071018_create_jenis_iurans_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id('jenis_iuran_id');
             $table->string('nama')->unique();
             $table->text('deskripsi')->nullable();
-            $table->decimal('jumlah', 15, 2);
+            $table->unsignedBigInteger('jumlah');
             $table->timestamps();
         });
     }
